Fix Image serving with http.

+++ demo_v1_User.h
@@ -16,7 +16,7 @@ namespace demo::v1 {
         //void get(const HttpRequestPtr& req,std::function<void (const HttpResponsePtr &)> &&callback,int p1,std::string p2);
         //void your_method_name(const HttpRequestPtr& req,std::function<void (const HttpResponsePtr &)> &&callback,double p1,int p2) const;
         void download(const HttpRequestPtr &req, std::function<void(const HttpResponsePtr &)> &&callback);
-        void download_id(const HttpRequestPtr &req, std::function<void(const HttpResponsePtr &)> &&callback, int id, int version);
-        void thumb_id(const HttpRequestPtr &req, std::function<void(const HttpResponsePtr &)> &&callback, int id, int version);
+        void download_id(const HttpRequestPtr &req, std::function<void(const HttpResponsePtr &)> &&callback, long id, int version);
+        void thumb_id(const HttpRequestPtr &req, std::function<void(const HttpResponsePtr &)> &&callback, long id, int version);
     };
 }
